Object Classification Project
"the object classification workflow aims to classify full objects, based on object-level features and user annotations. An object in this context is a set of pixels that belong to the same instance. In order to do so, the workflow needs segmentation images besides the usual raw image data, that can e.g. be generated with the Pixel Classification Workflow. " (Ilastik website)